feat: Apply contextmenu to DndTree as well
