+++ src/store/search.ts
@@ -7,6 +7,11 @@ interface SearchTriggerState {
   ) => void
 }
 
+interface SearchKeywordState {
+  term: string
+  setTerm: (updater: string | ((prev: string) => string)) => void
+}
+
 export const useSearchTriggerStore = create<SearchTriggerState>((set) => ({
   isSearchTriggered: false,
   setIsSearchTriggered: (updater) =>
@@ -14,5 +19,15 @@ export const useSearchTriggerStore = create<SearchTriggerState>((set) => ({
       typeof updater === 'function'
         ? { isSearchTriggered: updater(prev.isSearchTriggered) }
         : { isSearchTriggered: updater },
+    ),
+}))
+
+export const useSearchKeywordState = create<SearchKeywordState>((set) => ({
+  term: '',
+  setTerm: (updater) =>
+    set((prev) =>
+      typeof updater === 'function'
+        ? { term: updater(prev.term) }
+        : { term: updater },
     ),
 }))
